requirements.txt
gradio==5.9.1
coqui-tts==0.25.1
soundfile==0.12.1
transformers==4.33.0
streamlit
audio-recorder-streamlit
